refac(utils): reorganized calc utils
also updated imports in some files
