+++ src/utils/calc/index.ts
@@ -14,6 +14,7 @@ export * from './createSmogonMove';
 export * from './createSmogonPokemon';
 export * from './determineCriticalHit';
 export * from './determineMoveTargets';
+export * from './findHighestStat';
 export * from './getMoveOverrideDefaults';
 export * from './guessServerLegacySpread';
 export * from './guessServerSpread';
@@ -21,3 +22,4 @@ export * from './hasMoveOverrides';
 export * from './shouldBoostTeraStab';
 export * from './shouldIgnoreItem';
 export * from './statModRecorder';
+export * from './useSmogonMatchup';
